refactor StyledNavMenu as a seperate component
